Move parser helpers to irfmt::parsers
