+++ src/basic_block.rs
@@ -15,13 +15,16 @@ use crate::{
     error::Result,
     identifier::Identifier,
     indented_block,
-    irfmt::printers::{iter_with_sep, list_with_sep},
+    irfmt::{
+        parsers::{spaced, type_parser},
+        printers::{iter_with_sep, list_with_sep},
+    },
     linked_list::{private, ContainsLinkedList, LinkedList},
     location::{Located, Location},
     operation::Operation,
-    parsable::{self, spaced, IntoParseResult, Parsable, ParseResult},
+    parsable::{self, IntoParseResult, Parsable, ParseResult},
     printable::{self, indented_nl, ListSeparator, Printable},
-    r#type::{type_parser, TypeObj, Typed},
+    r#type::{TypeObj, Typed},
     region::Region,
     use_def_lists::{DefNode, Value},
 };
