+++ src/dialects/builtin/attributes.rs
@@ -15,11 +15,14 @@ use crate::{
     dialect::Dialect,
     error::Result,
     impl_attr_interface, input_err,
-    irfmt::printers::quoted,
+    irfmt::{
+        parsers::{spaced, type_parser},
+        printers::quoted,
+    },
     location::Located,
-    parsable::{spaced, IntoParseResult, Parsable, ParseResult, StateStream},
+    parsable::{IntoParseResult, Parsable, ParseResult, StateStream},
     printable::{self, Printable},
-    r#type::{type_parser, TypeObj, Typed},
+    r#type::{TypeObj, Typed},
     verify_err_noloc,
 };
 
@@ -485,7 +488,7 @@ mod tests {
     use expect_test::expect;
 
     use crate::{
-        attribute::{attr_cast, attr_parser},
+        attribute::attr_cast,
         context::Context,
         dialects::{
             self,
@@ -495,6 +498,7 @@ mod tests {
                 types::{IntegerType, Signedness},
             },
         },
+        irfmt::parsers::attr_parser,
         location,
         parsable::{self, state_stream_from_iterator},
         printable::Printable,
