+++ src/dialects/llvm/ops.rs
@@ -6,9 +6,10 @@ use crate::{
     error::Result,
     identifier::Identifier,
     impl_op_interface, input_err,
+    irfmt::parsers::ssa_opd_parser,
     location::{Located, Location},
     op::{Op, OpObj},
-    operation::{ssa_opd_parser, Operation},
+    operation::Operation,
     parsable::{Parsable, ParseResult},
     printable::{self, Printable},
     use_def_lists::Value,
