+++ src/irfmt/mod.rs
@@ -1,1 +1,2 @@
+pub mod parsers;
 pub mod printers;
